Added option to select matplotlib backend
